+++ CHANGELOG.rst
@@ -21,6 +21,7 @@ The format is based on `Keep a Changelog`_.
 Added
 -----
 - Improve efficiency of parallelization with max_memory_mb a new parameter of backend_opts (#61)
+- Add optimized mcx, mcy, mcz, mcu1, mcu2, mcu3, gates to QubitVector (#124)
 
 Changed
 -------